Fix MouseReleaseEvent not being sent when clicking on a context item.
The context menu that appears, when right clicking in a WebEngineView,
does not have a parent. On OSX if a context menu does not have a parent,
a MouseReleaseEvent is not sent when an item is chosen in the menu,
and thus consecutive mouse events are not handled properly.
The fix consists in assiging the WebEngineView as the QML parent
for the context menu that is opened.

Task-number: QTBUG-44666
Change-Id: I7f79d4497c57a814866089195da1570591a6999e
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>