BTUC-11019: Query auth credentialss from QNetworkProxy
If proxy authentication is required, username and password are now
queried from QNetworkProxy. If no valid credentials are available,
a dialog is shown for user to enter the credentials. This change was
required because at least on OS X several credential dialogs were shown
for user when loading a web page.

Change-Id: I3865c5343d5f2bb8a936cccbcc4ec0fab7e37334